make `NSURLSession` conform to `SessionDataTaskDataSource` by default
changes the `SessionDataTaskDataSource` protocol method of `func dataTaskWithRequest(request: NSURLRequest, completion: (NSData?, NSURLResponse?, NSError?) -> Void) -> NSURLSessionDataTask?` to be`func dataTaskWithRequest(request: NSURLRequest, completionHandler: (NSData?, NSURLResponse?, NSError?) -> Void) -> NSURLSessionDataTask` in order to enable direct conformance to `NSURLSession`
`
